requirements.txt
amqp==5.0.6
anyio==3.3.3
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
cachetools==4.2.4
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
debugpy==1.4.1
decorator==5.1.0
elasticsearch==7.16.1
entrypoints==0.3
fastapi==0.70.0
future==0.18.2
google-api-core==2.3.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
h11==0.12.0
idna==3.3
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.1
jupyter-core==4.8.1
kombu==5.1.0
libcst==0.3.23
matplotlib-inline==0.1.2
mixpanel==4.9.0
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
packaging==21.3
pandas==1.3.3
pandas-gbq==0.16.0
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.39
pickleshare==0.7.5
pip==21.2.4
prompt-toolkit==3.0.20
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybigquery==0.10.2
pydantic==1.8.2
pydata-google-auth==1.3.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
pyzmq==22.2.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
setuptools==58.0.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
starlette==0.16.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.0
email-validator
flower